Update the Amazon Ads Java Class file, now it uses a singleton class mode. Testing calling Java Objects from C++, using this we can call directly the methods from the Class itself and not from the Sonar Framworks Bridge. This is still a test in the Amazon Ads Java Class.
