+++ External Cocos Helper Android Frameworks/Frameworks/AdColony/AdColonyAds.java
