+++ External Cocos Helper Android Frameworks/Frameworks/Amazon/AmazonGameCircles.java
