+++ External Cocos Helper Android Frameworks/Frameworks/GooglePlayServices/GooglePlayServicesGameHelper.java
