+++ External Cocos Helper Android Frameworks/Frameworks/GooglePlayServices/GooglePlayServicesGameHelperUtils.java
