+++ External Cocos Helper Android Frameworks/proj.android-studio/app/src/sonar/systems/framework/SonarFrameworkSettings.java
@@ -31,7 +31,7 @@ public class SonarFrameworkSettings
 
     public static final boolean USE_AMAZON_GAMECIRCLES        = false;  // Requires Amazon GameCircle SDK
 
-    public static final boolean USE_AMAZON_ADS                = true; // Requires Amazon ads Jar file library
+    public static final boolean USE_AMAZON_ADS                = false; // Requires Amazon ads Jar file library
 
     public static final boolean USE_FLURRY_ANALYTICS          = false; // Requires Google play services & Flurry
                                                                        // Analytics Jar
