+++ External Cocos Helper Android Frameworks/proj.android/AndroidManifest.xml
@@ -131,7 +131,7 @@
       </intent-filter>
       </receiver> -->
        
-      <!-- End Requirements for use Amazon Game Circles DSK -->
+      <!-- End Requirements for use Amazon Game Circles SDK -->
 
       <!-- Requirements for Amazon Ads
